+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
  
 
 
+## [1.0-SNAPSHOT] - 2023-08-25
+ 
+### Added
+- Read vehicle names from JSON
+- Dropdown menu for selecting vehicle
+
+### Next
+- Read all data from selected vehicle
+- GUI development
+- Add data to JSON
+- Add and remove vehicles
+- Add events
+
+- Create new project
+- Save current JSON to user specified location
+
  
 ## [1.0-SNAPSHOT] - 2023-08-24